+++ src/backend/access/nbtree/nbtsort.c
@@ -482,9 +482,8 @@ _bt_buildadd(BTWriteState *wstate, BTPageState *state, IndexTuple itup)
 	if (itupsz > BTMaxItemSize(npage))
 		ereport(ERROR,
 				(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-			errmsg("index row size %lu exceeds maximum %lu for index \"%s\"",
-				   (unsigned long) itupsz,
-				   (unsigned long) BTMaxItemSize(npage),
+			errmsg("index row size %zu exceeds maximum %zu for index \"%s\"",
+				   itupsz, BTMaxItemSize(npage),
 				   RelationGetRelationName(wstate->index)),
 		errhint("Values larger than 1/3 of a buffer page cannot be indexed.\n"
 				"Consider a function index of an MD5 hash of the value, "
