+++ src/backend/storage/ipc/dsm.c
@@ -201,8 +201,8 @@ dsm_postmaster_startup(void)
 	dsm_control = dsm_control_address;
 	on_shmem_exit(dsm_postmaster_shutdown, 0);
 	elog(DEBUG2,
-		 "created dynamic shared memory control segment %u (%lu bytes)",
-		 dsm_control_handle, (unsigned long) segsize);
+		 "created dynamic shared memory control segment %u (%zu bytes)",
+		 dsm_control_handle, segsize);
 	dsm_write_state_file(dsm_control_handle);
 
 	/* Initialize control segment. */
