+++ src/backend/storage/ipc/dsm_impl.c
@@ -329,8 +329,8 @@ dsm_impl_posix(dsm_op op, dsm_handle handle, Size request_size,
 
 		ereport(elevel,
 				(errcode_for_dynamic_shared_memory(),
-				 errmsg("could not resize shared memory segment %s to %lu bytes: %m",
-					name, (unsigned long) request_size)));
+				 errmsg("could not resize shared memory segment %s to %zu bytes: %m",
+					name, request_size)));
 		return false;
 	}
 
@@ -871,8 +871,8 @@ dsm_impl_mmap(dsm_op op, dsm_handle handle, Size request_size,
 
 		ereport(elevel,
 				(errcode_for_dynamic_shared_memory(),
-				 errmsg("could not resize shared memory segment %s to %lu bytes: %m",
-					name, (unsigned long) request_size)));
+				 errmsg("could not resize shared memory segment %s to %zu bytes: %m",
+					name, request_size)));
 		return false;
 	}
 	else if (*mapped_size < request_size)
@@ -919,8 +919,8 @@ dsm_impl_mmap(dsm_op op, dsm_handle handle, Size request_size,
 
 			ereport(elevel,
 					(errcode_for_dynamic_shared_memory(),
-					 errmsg("could not resize shared memory segment %s to %lu bytes: %m",
-						name, (unsigned long) request_size)));
+					 errmsg("could not resize shared memory segment %s to %zu bytes: %m",
+						name, request_size)));
 			return false;
 		}
 	}
